UI: Fix nullable type navigation arguments (1/2)
